requirements.txt
dj-database-url==0.4.1
Django==1.9.8
gunicorn==19.6.0
psycopg2==2.6.2
whitenoise==3.2
django-compressor==1.5
django-debug-toolbar==1.3.2
django-extensions==1.5.5
ipython==3.2.0
six==1.9.0
sqlparse==0.1.15
wsgiref==0.1.2
bootstrap_admin==0.3.7.1
django-admin-tools==0.8.0
django-admin-tools-stats==0.7.1
django-import-export==0.5.1
django-autocomplete-light==3.2.0
tablib
django-calendarium
django-daterange-filter
